+++ docs/source/changelog.rst
@@ -5,6 +5,8 @@ Change Log
 **Pyro 4.39**
 
 - dropped support for Python 2.6 and Python 3.2. Supported versions are now 2.7, 3.3, 3.4 and up.
+- better exception when message size exceeds 2 gigabyte limit
+- mentioned the 2 gigabyte message size limit in the docs